+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Fixed
 - Japanese in english locales for Gotoaddress, addclass dialog
+- AS1/2 DefineFunction cleaner
 
 ## [14.5.2] - 2021-11-20
 ### Fixed